application/requirements.txt
django
django-request-id
django-request-logging
django-rest-framework
drf-yasg
psycopg2-binary
